requirements.txt
torch==2.1.0
transformers==4.35.2
PyYAML==6.0
matplotlib==3.7.0
numpy==1.26.4
tqdm==4.66.1